# Improve Evib/Erot fetch performances