### Description
<!-- Provide a general description of what your pull request does. -->

- Evib/Erot was always fetched in Treanor distributions, even if already in the database. Fixed. --> **CO2 3-Tvib calculations can be up to 5x faster**  @BlehMaks  @CorentinGrimaldi 

- implemented index.map(dict.get) method to fetch Evib/Erot for  (CO2  in 3-T Treanor distribution) and (CO non-equilibrium 
  --> **fetching rovibrational energies  is about ~30% faster** than groupby().apply() 

---

### Details 

Test case : 
```python
from radis import SpectrumFactory
sf = SpectrumFactory(2284, 2285,         # cm-1
                  molecule='CO2',
                  isotope='1,2,3',
                  pressure=1.01325,   # bar
                  mole_fraction=0.1,
                  path_length=1,      # cm
                  wstep=0.001,
                  verbose=3)
sf.load_databank("HITEMP-CO2-DUNHAM")

```
First calculation with vib_distribution='boltzmann'
```python
sf.non_eq_spectrum(Tvib=(900, 900, 1100), Trot=700,
                    vib_distribution='boltzmann')
"""
     ... Fetched energies in 1.56s for all 32763 transitions
     ... calculated weighted transition moment in 0.11s
     Checked nonequilibrium parameters in 1.77s
"""
```
2nd calculation : 

```python
sf.non_eq_spectrum(Tvib=(900, 900, 1100), Trot=700,
                    vib_distribution='boltzmann')
"""
     Checked nonequilibrium parameters in 0.00s
"""
```
( Evib Erot are not re-fetched, as expected !)  ✔️ 


In Treanor things were diffferent : 

First calculation : 

```python
sf.non_eq_spectrum(Tvib=(900, 900, 1100), Trot=700,
                    vib_distribution='treanor')

"""
    ... Fetched energies in 2.40s for all 32763 transitions
    ... calculated weighted transition moment in 0.11s
    Checked nonequilibrium parameters in 2.63s
"""
```

2nd calculation : 
```python
sf.non_eq_spectrum(Tvib=(900, 900, 1100), Trot=700,
                    vib_distribution='treanor')

'''
     ... Fetched energies in 2.11s for all 32763 transitions
     Checked nonequilibrium parameters in 2.11s
'''
```
no change    ⚠️   


**This is fixed now.** 

Total calculation time  for 2nd calculation ::
```
    Spectrum calculated in 2.51s
```
becomes :
```
    Spectrum calculated in 0.52s
```

---

2nd change : the fetch function itself was improved, by replacing apply.group() by index.map(dict.get()) 

```python
"""
    ... Fetched energies in 1.72s for all 32763 transitions
    ... calculated weighted transition moment in 0.11s
    Checked nonequilibrium parameters in 1.94s
"""
```
 (compared to 2.4s for Fetched energy : ~30% faster)

---


Fixes #193